[dqbe] Replace size_t with int in app.d
This avoids compilation failures in the bootstrap compiler, which only supports int, char, and bool as primitive types.

+++ source/dqbe/app.d
@@ -10,16 +10,16 @@ import dqbe.codegen;
 extern (C)
 int main(int argc, char** argv) {
   // Read all inputs from stdin
-  size_t capacity = 4 * 1024 * 1024; // 4MB starting capacity
+  int capacity = 4 * 1024 * 1024; // 4MB starting capacity
   char* buf = cast(char*) malloc(capacity);
   if (!buf) {
     fprintf(stderr, "Out of memory allocating input buffer\n");
     return 1;
   }
   
-  size_t len = 0;
+  int len = 0;
   while (true) {
-    size_t read_bytes = fread(buf + len, 1, 4096, stdin);
+    int read_bytes = fread(buf + len, 1, 4096, stdin);
     if (read_bytes == 0) {
       break;
     }
